sandbox: reorgize files
