--- internal/sandbox/network_darwin.go
@@ -1,17 +1,0 @@
-package sandbox
-
-import "golang.org/x/sys/unix"
-
-func setCloseOnExecAndNonBlocking(fd int) error {
-	if _, err := ignoreEINTR2(func() (int, error) {
-		return unix.FcntlInt(uintptr(fd), unix.F_SETFD, unix.O_CLOEXEC)
-	}); err != nil {
-		return err
-	}
-	if err := ignoreEINTR(func() error {
-		return unix.SetNonblock(fd, true)
-	}); err != nil {
-		return err
-	}
-	return nil
-}
