+++ internal/sandbox/socket_unix.go
@@ -1,9 +1,9 @@
 package sandbox
 
-func (s *socketFD) release(fd int) {
-	s.releaseFunc(fd, closeTraceError)
+func (f *fdRef) release(fd int) {
+	f.releaseFunc(fd, closeTraceError)
 }
 
-func (s *socketFD) close() {
-	s.closeFunc(closeTraceError)
+func (f *fdRef) close() {
+	f.closeFunc(closeTraceError)
 }
